@@ -17,41 +17,48 @@ buildscript {
 apply plugin: 'java'
 apply plugin: 'eclipse'
 apply plugin: 'idea'
-apply plugin: 'spring-boot'
+apply plugin: 'org.springframework.boot'
 apply plugin: 'io.spring.dependency-management'
 apply plugin: 'org.sonarqube'
+apply plugin: 'jacoco'
+
 
 jar {
-    baseName = 'outout'
+    //baseName = 'outout'
     version = '0.0.1-SNAPSHOT'
 }
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
-def springfoxVersion = '2.6.1'
-def springfoxUiVersion = '2.6.1'
+sourceCompatibility = 21
+targetCompatibility = 21
+def springfoxVersion = '3.0.0'
+def springfoxUiVersion = '3.0.0'
 repositories {
     mavenCentral()
 }
 
-
 dependencies {
-    compile("io.springfox:springfox-swagger2:${springfoxVersion}")
-    compile("io.springfox:springfox-swagger-ui:${springfoxUiVersion}")
-    compile("javax.servlet:javax.servlet-api:3.1.0")
-    compile("org.springframework.boot:spring-boot-starter-web")
-    compile("org.springframework.boot:spring-boot-starter-data-jpa")
-    compile("org.springframework.boot:spring-boot-starter-security")
-    compile 'io.jsonwebtoken:jjwt:0.5.1'
-    compile 'joda-time:joda-time:2.3'
-    compile("com.h2database:h2")
-    testCompile("org.springframework.boot:spring-boot-starter-test")
-    testCompile("junit:junit:4.12")
-    testCompile("org.hamcrest:hamcrest-all:1.3")
-    testCompile("org.apache.tomcat:tomcat-catalina:8.0.24")
-    testCompile("org.apache.commons:commons-dbcp2:2.1.1")
-    testCompile("org.exparity:hamcrest-date:1.1.0")
+    implementation 'joda-time:joda-time:2.3'
 
+    implementation 'org.springframework.boot:spring-boot-starter-validation:3.0.0'
+    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.4.0'
+    implementation("javax.servlet:javax.servlet-api:4.0.1")
+    implementation("org.springframework.boot:spring-boot-starter-web:${springboot_version}")
+    implementation("org.springframework.boot:spring-boot-starter-data-jpa:${springboot_version}")
+    implementation("org.springframework.boot:spring-boot-starter-security:${springboot_version}")
+    implementation 'org.springframework.security:spring-security-web:6.2.3'
+    implementation 'org.springframework.security:spring-security-config:6.2.3'
+    implementation 'io.jsonwebtoken:jjwt:0.12.5'
+    implementation 'org.joda:joda-money:1.0.3'
+    implementation("com.h2database:h2:1.4.192")
+    compileOnly 'jakarta.servlet:jakarta.servlet-api:6.1.0-M2'
+    testImplementation("org.springframework.boot:spring-boot-starter-test:${springboot_version}")
+    testImplementation("junit:junit:4.13.2")
+    testImplementation("org.hamcrest:hamcrest-all:1.3")
+    implementation 'org.springframework.boot:spring-boot-starter-tomcat:3.2.4'
+    testImplementation("org.apache.commons:commons-dbcp2:2.9.0")
+    testImplementation("org.exparity:hamcrest-date:2.0.0")
+    testImplementation 'org.mockito:mockito-core:4.2.0'
 }
+
 sonarqube {
     properties {
         property "sonar.projectKey", "andres-marcelo-csdt-project_outout-api"
