@@ -1,10 +1,14 @@
 package outout;
 
+
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import org.h2.server.web.WebServlet;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.embedded.ServletRegistrationBean;
+
+import org.springframework.boot.web.servlet.ServletRegistrationBean;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.core.env.Environment;
 import org.springframework.orm.jpa.JpaTransactionManager;
@@ -12,8 +16,11 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.crypto.password.StandardPasswordEncoder;
 import org.springframework.transaction.PlatformTransactionManager;
 
+import jakarta.servlet.Servlet;
+
 @SpringBootApplication
-public class OutoutApplication {
+@OpenAPIDefinition
+public class OutoutApplication extends SpringBootServletInitializer {
 
     @Autowired
     private Environment environment;
@@ -32,11 +39,11 @@ public class OutoutApplication {
         return new StandardPasswordEncoder(environment.getProperty("password.encoder.secret"));
     }
 
-    @Bean
-    public ServletRegistrationBean h2servletRegistration() {
-        ServletRegistrationBean registration = new ServletRegistrationBean(new WebServlet());
-        registration.addUrlMappings("/console/*");
-        return registration;
-    }
+//    @Bean
+//    public ServletRegistrationBean h2servletRegistration() {
+//        ServletRegistrationBean registration = new ServletRegistrationBean(new WebServlet());
+//        registration.addUrlMappings("/console/*");
+//        return registration;
+//    }
 
 }
