@@ -1,41 +1,54 @@
 package outout.config;
 
+import io.swagger.v3.oas.models.Components;
+import io.swagger.v3.oas.models.ExternalDocumentation;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.info.License;
+import io.swagger.v3.oas.models.parameters.Parameter;
+import io.swagger.v3.oas.models.security.SecurityScheme;
+import org.springdoc.core.customizers.OperationCustomizer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import springfox.documentation.builders.ParameterBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.schema.ModelRef;
-import springfox.documentation.service.Parameter;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
+//import springfox.documentation.builders.ParameterBuilder;
+//import springfox.documentation.builders.PathSelectors;
+//import springfox.documentation.builders.RequestHandlerSelectors;
+//import springfox.documentation.schema.ModelRef;
+//import springfox.documentation.service.Parameter;
+//import springfox.documentation.spi.DocumentationType;
+//import springfox.documentation.spring.web.plugins.Docket;
+//import springfox.documentation.swagger2.annotations.EnableSwagger2;
 import java.util.*;
 
 @Configuration
-@EnableSwagger2
+//@EnableSwagger2
 
 public class SwaggerConfig {
     private static final Set<String> DEFAULT_PRODUCES_CONSUMES = new HashSet<String>(Arrays.asList("application/json"));
 
     @Bean
-    public Docket api() {
-        ParameterBuilder parameterBuilder = new ParameterBuilder();
-        parameterBuilder.name("X-AUTH-TOKEN")
-                .modelRef(new ModelRef("string"))
-                .parameterType("header")
-                .description("JWT token")
-                .required(false)
-                .build();
-        List<Parameter> parameters = new ArrayList<>();
-        parameters.add(parameterBuilder.build());
-        return new Docket(DocumentationType.SWAGGER_2)
-                .produces(DEFAULT_PRODUCES_CONSUMES)
-                .consumes(DEFAULT_PRODUCES_CONSUMES)
-                .select()
-                .build()
-                // Setting globalOperationParameters ensures that authentication header is applied to all APIs
-                .globalOperationParameters(parameters);
+    public OpenAPI springShopOpenAPI() {
+        return new OpenAPI()
+                .info(new Info().title("outout API")
+                        .description("csdt project application")
+                        .version("v0.0.1")
+                        .license(new License().name("Apache 2.0").url("http://springdoc.org")))
+                .externalDocs(new ExternalDocumentation()
+                        .description("SpringShop Wiki Documentation")
+                        .url("https://springshop.wiki.github.org/docs"));
+
+        //add security header to all endpoints
+
+    }
+
+    @Bean
+    public OperationCustomizer customize() {
+        return (operation, handlerMethod) -> operation.addParametersItem(
+                new Parameter()
+                        .in("header")
+                        .required(false)
+                        .description("Authorization Header")
+                        .name("X-AUTH-TOKEN"));
     }
 
 }
